Rationalised file structure, deleted old castep phonon parser
